Interface.Interaction

Description

This module contains functions that process input from command-line and output the result. It is used by Main.

Documentation

printParliament :: Parliament -> IO ()

Outputs the parliament in a csv format with columns "party,nubmer_of_seats".

readInput

Arguments

:: String
file with districts

-> String
file with voters

-> Int
number of parties

-> IO Input

Reads VirtualArrow.Input.Input data type from supplied csv files. Used by runResultCommand and runRCommand.

run :: Command -> IO ()

Read command-line arguments and starts processing of the supplied command. Used by Main in the main function.