VirtualArrow.Election

Description

Documentation

bordaCount :: Input -> Parliament

Each party gets a number of points for each ballot: 0 if it is on the first place in the ballot, 1 if on the second and so on till n - 1 points. A candidate with the least number of votes wins. To determine the resulting parliament find a winner for each district. Assign all seats to the winner, sum seats across districts.

oneDistrictProportionality :: Input -> Parliament

Districts are aggregated into one and seats in the parliament are distributed proportionally to the first preferences in the electorate.

plurality :: Input -> Parliament

The party with the most votes wins in each district.

runOffPlurality :: Input -> Parliament

In each district all parties but the two with the most votes are excluded. The second round is implemented with these two parties only and the one with the most votes wins. If after the first round the first party has at least 50% of the votes, it is elected without the need of a second round.

multiDistrictProportionality :: Input -> Parliament

In each district seats are allocated using Sainte-Laguë method: successive quotients are calculated for each party, whichever party has the highest quotient gets the next seat allocated. The process is repeated until all seats have been allocated.

mixedMember1 :: Input -> Double -> Parliament

One parliament is elected with the Plurality System, and one with Proportional System. The resulting parliament is a weighted mean of the two. Weight of the first parliament relative to the second is specified.

mixedMember2 :: Input -> Double -> Parliament

Part of parliament is elected with the Plurality System, and the remainder is elected using the Proportional System. Share of seats to be elected with the Plurality System is specified.

thresholdProportionality :: Input -> Double -> Parliament

All the parties who have an overall share of votes (strictly) smaller are excluded from the parliament. The seats are distributed proportionally among the remaining parties. There is only one district.

singleTransferableVote

Arguments

:: Input

-> Map Int Int
keys are candidate ids and values are their parties

-> Parliament

The seats for each party in each district are assigned according to a quota value. If some seats are not assigned, the votes unused by the elected candidates are transferred to the next candidate in the elector's preference ordering, and the candidates with the highest number of votes are elected. The operation is repeated until all the seats have been assigned. If at a given round there is no assignment, the candidate with less preference is excluded, and its votes are distributed as above. The Droop quota is used and the votes are redistributed using the Gregory method (transfers partial votes). Single Transferable Vote is used with the list of candidates, thus the additional parameter (map) is required to determine the number of seats for each party.