VirtualArrow.Input

Description

This module defines data structures and accompanying functions.

Types

type Preferences = Vector Int

Vector of parties ordered by voter's preference.

type DistrictID = Int

type Party = Int

type NumberOfSeats = Int

type Parliament = [(Party, NumberOfSeats)]

The resulting parliament.

Data types

data Candidate

Constructors

Candidate

Fields

candidateID :: !Int

party :: !Int

data District

Represents one electoral district (constituency).

Constructors

District

Fields

districtID :: !DistrictID

nseats :: !NumberOfSeats

data Voter

Represents one voter in a district.

Constructors

Voter

Fields

district :: !DistrictID

preferences :: !Preferences

data Input

Represents input to the program - collection of districts, voters number of parties and internally calculated (using votersByDistrict) map from district id to votes.

Constructors

Input

Fields

districts :: ![District]

voters :: ![Voter]

nparties :: !Int

districtMap :: Map DistrictID [Voter]

Functions

parliamentSize :: Input -> NumberOfSeats

Total number of seats.

numberOfSeatsByDistrictID :: Input -> DistrictID -> NumberOfSeats

Returns number of seats by district id.

numberOfSeatsByDistrict :: Input -> [(DistrictID, NumberOfSeats)]

nvoters :: Input -> Int

Total number of voters.

firstChoices :: Input -> [Party]

Returns list of first choices (first preference) for each voter in the input.

firstChoicesAmongVoters :: [Voter] -> [Party]

Returns list of first choices (first preference) for each voter in the list.

calculateProportion

Arguments

:: Input

-> Int
number of votes

-> Int

Returns proportion of parliament corresponding to the number of votes.

prefToPlaces :: Preferences -> Vector Int

Returns vector of places, s.t. ! i is the place of ith party in the list of preferences.

votersByDistrict :: [Voter] -> [(DistrictID, [Voter])]

candidateMap :: [Candidate] -> Map Int Party

Returns map from candidate id to party.